Version 2

+++ app/Ajax.php
@@ -12,31 +12,47 @@ class Ajax
 
     private static function add_to_favorites()
     {
-        Info::get('db')->exec("insert into favorites (SONGID) values (".mysqli_escape_string(Info::get('dbh'), self::$args['id']).")");
+        Info::get('db')->exec("insert into favorites (groupId, SONGID) values ({$_SESSION['userId']},".mysqli_escape_string(Info::get('dbh'), self::$args['id']).")");
         return '';
     }
 
     private static function add_to_piano_favorites()
     {
-        Info::get('db')->exec("insert into piano_favorites (SONGID) values (".mysqli_escape_string(Info::get('dbh'), self::$args['id']).")");
+        Info::get('db')->exec("insert into piano_favorites (groupId, SONGID) values ({$_SESSION['userId']},".mysqli_escape_string(Info::get('dbh'), self::$args['id']).")");
         return '';
     }
+
 
     private static function get_favorites()
     {
         $sql = "SELECT f.ID as FID, l.*, concat(l.num, ' - ',l.name) as dispName, 
-                        concat('/images/',l.LISTID,'/',l.num,'.jpg') as imageName FROM favorites f 
-                left join song_list l ON l.ID=f.SONGID 
+                        concat('/images/',l.LISTID,'/',l.num,'.jpg') as imageName, f.SONGID FROM favorites f 
+                left join song_list l ON l.ID=f.SONGID
+                where f.groupId={$_SESSION['userId']}                                                                        
                 ORDER BY FID";
         $list = Info::get('db')->select($sql);
         return json_encode($list);
     }
 
+
+    private static function get_favorites_with_text()
+    {
+        $sql = "SELECT f.ID as FID, l.*, concat(l.num, ' - ',l.name) as dispName, 
+                        concat('/images/',l.LISTID,'/',l.num,'.jpg') as imageName, f.SONGID, l.TEXT FROM favorites f 
+                left join song_list l ON l.ID=f.SONGID
+                where f.groupId={$_SESSION['userId']}                                                                        
+                ORDER BY FID";
+        $list = Info::get('db')->select($sql);
+        return json_encode($list);
+    }
+
+
     private static function get_piano_favorites()
     {
         $sql = "SELECT f.ID as FID, l.*, concat(l.num, ' - ',l.name) as dispName, 
-                        concat('/images/',l.LISTID,'/',l.num,'.jpg') as imageName FROM piano_favorites f 
-                left join song_list l ON l.ID=f.SONGID 
+                        concat('/images/',l.LISTID,'/',l.num,'.jpg') as imageName, f.SONGID FROM piano_favorites f 
+                left join song_list l ON l.ID=f.SONGID
+                where f.groupId={$_SESSION['userId']}                                                                  
                 ORDER BY FID";
         $list = Info::get('db')->select($sql);
         return json_encode($list);
@@ -44,14 +60,14 @@ class Ajax
 
     private static function clear_favorites()
     {
-        $sql = "DELETE FROM favorites";
+        $sql = "DELETE FROM favorites WHERE groupId={$_SESSION['userId']}";
         Info::get('db')->exec($sql);
         return '';
     }
 
     private static function clear_piano_favorites()
     {
-        $sql = "DELETE FROM piano_favorites";
+        $sql = "DELETE FROM piano_favorites WHERE groupId={$_SESSION['userId']}";
         Info::get('db')->exec($sql);
         return '';
     }
@@ -74,22 +90,45 @@ class Ajax
 
     private static function set_image()
     {
-        Info::get('db')->exec("insert into current (image) values ('/images/".
+        $txt = Info::get('db')->select("select TEXT from song_list where ID=".self::$args['song_id']);
+        Info::get('db')->exec("insert into current (groupId, image, text) values ({$_SESSION['userId']}, '/images/".
             mysqli_escape_string(Info::get('dbh'), self::$args['list_id'])."/".
-            mysqli_escape_string(Info::get('dbh'), self::$args['image_num']).".jpg')");
+            mysqli_escape_string(Info::get('dbh'), self::$args['image_num']).".jpg', \"{$txt[0]['TEXT']}\")");
         return '';
     }
 
     private static function get_image()
     {
-        $img = Info::get('db')->select("select image from current");
+        $img = Info::get('db')->select("select image from current where groupId=".$_SESSION['userId']);
         return json_encode($img);
+    }
+
+    private static function get_text()
+    {
+        $txt = Info::get('db')->select("select text from current where groupId=".$_SESSION['userId']);
+        return json_encode($txt);
+    }
+
+    private static function get_whole_text()
+    {
+        $songId = mysqli_escape_string(Info::get('dbh'), self::$args['id']);
+        $txt = Info::get('db')->select("select TEXT from song_list where ID={$songId}");
+        return json_encode($txt[0]);
+    }
+
+    private static function set_text()
+    {
+        $text = mysqli_escape_string(Info::get('dbh'), self::$args['text']);
+        $image = "/images/".mysqli_escape_string(Info::get('dbh'), self::$args['list_id'])."/".
+            mysqli_escape_string(Info::get('dbh'), self::$args['image_num']).".jpg";
+        Info::get('db')->exec("update current set text=\"{$text}\" WHERE groupId={$_SESSION['userId']} and image=\"{$image}\"");
+        return '';
     }
 
 
     private static function clear_image()
     {
-        Info::get('db')->exec("delete from current");
+        Info::get('db')->exec("delete from current where groupId=".$_SESSION['userId']);
         return '';
     }
 
@@ -100,6 +139,10 @@ class Ajax
     public static function execute($cmd)
     {
         $command = $cmd['command'];
+        if( !isset($_SESSION['userId']) ){
+            return json_encode(array('status'=>false, 'message'=>'User not logged in!'));
+        }
+
         if (is_callable(array('Ajax', $command))){
             self::$args = $cmd;
             $data = self::$command();
